Switch route53 over to the standard client auth setup
Initialize the client via the helper in EC2 which allows for the various auth methods that weren't available previously. This required a few changes to things. First off aws_region as the helper name for the aws region conflicted with the method alias used for backwards compatibility and caused a horrible loop. That helper has been renamed to avoid the loop since the method alias cannot be renamed. The client initialization helper has also been updated to use Ruby's double splat (**) to accept zero or more hashes which can be accessed with an options hash within the method. This was necessary so that route53 could continue to pass the "mock" option. We might want to set this up in other resources later on and now we can since setting mock: true will skip the auth and instead mock the resource in the AWS SDK.

Signed-off-by: Tim Smith <tsmith@chef.io>