locking aws gems to their latest minor version
this should make it so that only newly released patches for the respective aws gems are installed by default. This is because aws has a bunch of gems that release daily, this causes hundreds of versions of the same gems to be installed. In our scenerio at work just including the cookbook and making a single ssm call resulted in installing hundreds of gems (which some of the versions left it non functional and looking for shared object files that did not exist. We tried rebuilding the gems via `--pristine` and only after going a `gem --cleanup` did we see any improvement. After doing a cleaned our converge time went from `6 min 04 sec` to `18 sec`.

Signed-off-by: Ben Abrams <me@benabrams.it>