Add template example back in
This was removed in 337c45a8a4d4d35d64728ef6a47abd0fb9186148, but the
commit does not explain why. The PR seems to suggest a change in nested
hashes, but does that change accessing behaviour in run_state?

When coming to use this resource, I wasn't aware of the `lazy`
evaluation (I guess I don't know Chef as well I could), and spent some time
figuring out how to put values into a template. I only found this
example had been removed after I'd figured it out.

I don't think it hurts to have the example in the README. If it's
now incorrect, can we amend it instead of completely removing it?